src/requirements.txt
aiofiles==0.8.0
blinker==1.4
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
google-api-core==2.7.2
google-api-python-client==2.46.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.0
grpcio==1.44.0
grpcio-tools==1.44.0
h11==0.13.0
h2==4.1.0
hpack==4.0.0
httplib2==0.20.4
hypercorn==0.13.2
hyperframe==6.0.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
priority==2.0.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.8
quart==0.17.0
requests==2.27.1
rsa==4.8
six==1.16.0
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.1.1
wsproto==1.1.0
